Enable auto formatting
This will introduce both Black and isort. As part of adding Black, a few
of the previous Flake8 plugins are being removed and Bugbear is being
added.

Most of the current lint settings are being left alone -- the most
noticeable exception is line length, which is being reduced by two
characters to match Black's default -- while being moved to `setup.cfg`
to make them accessible outside of tox.

Closes #49