Fix local development email settings
Mailhog does not support TLS, so the hardcoded setting for
MAIL_USE_TLS=True breaks local development. This change makes it work
again.

This also adds the code as a volume in the compose file to remove the
need to restart the compose setup if the only changes are code-related.