Publish Docker images during the pipeline
In an effort to speed up the `docker` job, this will publish an image of
`master` under Docker's `ci` tag. This image will hopefully save us time
in future builds, especially ones that don't introduce new requirements.
Both this new job and the existing `docker` job are being placed in a
new stage of the pipeline that will only run if the other jobs pass.

Closes #51