Use Postgres for tests
This will switch from SQLite to Postgres for all tests. Given the
current structure of the settings, there isn't a great way to provide
both local and CI test settings. The settings here are based on CI and
references to running the tests locally are being removed or annotated
to mention that.