Don't parse email subjects as YAML
We need to parse the rendered email templates to extract the
frontmatter. This means that certain characters (e.g., `:`) will cause
the parsing to fail. By wrapping the values in quotes, the parser will
treat the entire value as a string.

Note that including a double quote in any of the context values included
in these strings (currently all subjects) can cause a similar error.

Fixes YAK-BAK-H