Make co-presenter email subject more explicit
The email sent to people added as co-presenters can be interpreted as an
actual invitation to speak. This will make it clearer that the
invitation is to join as a co-presenter of a proposed talk, not an
accepted one.

The template itself is also being renamed to be more consistent with
other wording used.