Support clearing skipped votes
As a voter is progressing through the voting queue, they can currently
skip votes. The current behavior of this function is to add the votes to
the history table and do nothing else. This new feature allows a voter
to clear their skipped votes and continue voting a queue-like fashion as
opposed to selecting individual skipped talks to continue voting on. A
voter can clear their skipped votes at any time using the "Clear skipped
votes" button from their voting dashboard, and skipped votes will
automatically be cleared upon completion of a category. The "talks
remaining" count for each category on the dashboard becomes more clear
with this change; it never counted skipped votes, but now a user will
not end up with an empty voting queue and a positive talks remaining
count.